move Events to an own class
